Extract configuration to the separate class, adjust test
